hero and header
connects the header and hero section.
all paths are configured.
