+++ src/aboutUs.html
@@ -15,6 +15,7 @@
     <script type="module" src="./index.js"></script>
   </head>
   <body>
+    <include src="./partials/header.html"></include>
     <main>
       <include src="./partials/breadcrumbs.html"></include>
       <include src="./partials/about.html"></include>
