+++ src/design.html
@@ -10,14 +10,15 @@
       rel="stylesheet"
     />
 
-    <title>pr</title>
-    <link rel="stylesheet" href="../sass/index.scss" />
-    <script type="module" src="../index.js"></script>
+    <title>design</title>
+    <link rel="stylesheet" href="./sass/index.scss" />
+    <script type="module" src="./index.js"></script>
   </head>
   <body>
+    <include src="./partials/header.html"></include>
     <main>
       <include src="./partials/breadcrumbs.html"></include>
-      <include src="./partials/services/prInfo.html"></include>
+      <include src="./partials/services/designInfo.html"></include>
     </main>
 
     <include src="./partials/footer.html"></include>
