+++ src/digital.html
@@ -11,10 +11,11 @@
     />
 
     <title>digital</title>
-    <link rel="stylesheet" href="../sass/index.scss" />
-    <script type="module" src="../index.js"></script>
+    <link rel="stylesheet" href="./sass/index.scss" />
+    <script type="module" src="./index.js"></script>
   </head>
   <body>
+    <include src="./partials/header.html"></include>
     <main>
       <include src="./partials/breadcrumbs.html"></include>
       <include src="./partials/services/digitalInfo.html"></include>
