+++ src/partials/header.html
@@ -2,7 +2,7 @@
   <div class="container header__container">
     <div class="logo__container">
       <a
-        href="./index.html"
+        href="./"
         title="Головнa сторінка"
         rel="noopener noindex nofollow"
         class="logo__link"
@@ -34,14 +34,18 @@
       <nav>
         <ul>
           <li>
-            <a href="#home" title="Home" class="site-nav__link">Головнa</a>
+            <a href="./" title="Home" class="site-nav__link">Головнa</a>
           </li>
           <li>
-            <a href="#home" title="Home" class="site-nav__link">Про компанію</a>
+            <a href="./aboutUs.html" title="Home" class="site-nav__link"
+              >Про компанію</a
+            >
           </li>
           <li>
             <div class="site-nav__item-container">
-              <a href="#home" title="Home" class="site-nav__link">Послуги</a>
+              <a href="./services.html" title="Home" class="site-nav__link"
+                >Послуги</a
+              >
               <button
                 ria-label="відкрити меню послуг"
                 type="button"
@@ -63,34 +67,45 @@
               </button>
             </div>
 
-            <ul>
+            <ul class="site-nav__dropdown">
               <li class="site-nav__item-services">
-                <a href="#home" title="Home" class="site-nav__link-services"
+                <a href="./pr.html" title="Home" class="site-nav__link-services"
                   >PR</a
                 >
               </li>
               <li class="site-nav__item-services">
-                <a href="#home" title="Home" class="site-nav__link-services"
+                <a
+                  href="./design.html"
+                  title="Home"
+                  class="site-nav__link-services"
                   >Design</a
                 >
               </li>
               <li class="site-nav__item-services">
-                <a href="#home" title="Home" class="site-nav__link-services"
+                <a
+                  href="./smm.html"
+                  title="Home"
+                  class="site-nav__link-services"
                   >SMM</a
                 >
               </li>
               <li class="site-nav__item-services">
-                <a href="#home" title="Home" class="site-nav__link-services"
+                <a
+                  href="./digital.html"
+                  title="Home"
+                  class="site-nav__link-services"
                   >Digital</a
                 >
               </li>
             </ul>
           </li>
           <li>
-            <a href="#home" title="Home" class="site-nav__link">Вакансії</a>
+            <a href="./vacancies.html" title="Home" class="site-nav__link"
+              >Вакансії</a
+            >
           </li>
           <li>
-            <a href="#home" title="Home" class="site-nav__link"
+            <a href="./testimonials.html" title="Home" class="site-nav__link"
               >Відгуки про нас</a
             >
           </li>
