+++ src/partials/hero.html
@@ -10,22 +10,22 @@
       <li class="hero__nav-item">
         <div class="hero__nav-circle"></div>
         <div class="hero__nav-line"></div>
-        <a class="hero__nav-link" href="">PR</a>
+        <a class="hero__nav-link" href="./pr.html">PR</a>
       </li>
       <li class="hero__nav-item">
         <div class="hero__nav-circle"></div>
         <div class="hero__nav-line"></div>
-        <a class="hero__nav-link" href="">Design</a>
+        <a class="hero__nav-link" href="./design.html">Design</a>
       </li>
       <li class="hero__nav-item">
         <div class="hero__nav-circle"></div>
         <div class="hero__nav-line"></div>
-        <a class="hero__nav-link" href="">SMM</a>
+        <a class="hero__nav-link" href="./smm.html">SMM</a>
       </li>
       <li class="hero__nav-item">
         <div class="hero__nav-circle"></div>
         <div class="hero__nav-line"></div>
-        <a class="hero__nav-link" href="">Digital</a>
+        <a class="hero__nav-link" href="./digital.html">Digital</a>
       </li>
     </ul>
   </div>
