+++ src/sass/_hero.scss
@@ -9,9 +9,7 @@
 }
 
 .hero__content {
-    padding-top: 250px;
-    padding-left: 5%;
-    padding-right: 5%;
+    padding: 250px 5% 0;
 }
 
 .hero__title {
@@ -36,22 +34,19 @@
     display: flex;
     justify-content: center;
     background-color: var(--main-red-color);
-    padding-top: 28px;
-    padding-bottom: 28px;
+    padding: 28px 0;
     margin-bottom: 30px;
     border-radius: 260px;
     border: none;
     color: var(--main-title-color);
     font-size: 16px;
     font-weight: 600;
+    text-align: center;
 
     &:hover {
         background-color: transparent;
-        transition-delay: 0s;
-        transition-duration: 0.3s;
-        transition-property: all;
-        transition-timing-function: ease;
-        box-shadow: var(--main-red-color) 0px 0px 0px 2px inset;
+        transition: all 0.3s ease;
+        box-shadow: var(--main-red-color) 0 0 0 2px inset;
     }
 }
 
@@ -59,7 +54,7 @@
     list-style: none;
     padding: 0;
     display: grid;
-    grid-template-columns: repeat(1, 1fr);
+    grid-template-columns: 1fr;
     grid-row-gap: 10px;
 }
 
@@ -86,7 +81,7 @@
     display: flex;
     align-items: center;
     border-radius: 16px;
-    box-shadow: rgba(255, 255, 255, 0.4) 0px 0px 0px 1px inset;
+    box-shadow: rgba(255, 255, 255, 0.4) 0 0 0 1px inset;
     width: 100%;
     height: 75.5px;
     padding-left: 24px;
@@ -99,6 +94,6 @@
     &:hover {
         background-color: var(--main-title-color);
         color: var(--second-main-text-color);
-        transition: all 0.3s ease 0s;
+        transition: all 0.3s ease;
     }
 }
