+++ src/smm.html
@@ -10,14 +10,15 @@
       rel="stylesheet"
     />
 
-    <title>design</title>
-    <link rel="stylesheet" href="../sass/index.scss" />
-    <script type="module" src="../index.js"></script>
+    <title>smm</title>
+    <link rel="stylesheet" href="./sass/index.scss" />
+    <script type="module" src="./index.js"></script>
   </head>
   <body>
+    <include src="./partials/header.html"></include>
     <main>
       <include src="./partials/breadcrumbs.html"></include>
-      <include src="./partials/services/designInfo.html"></include>
+      <include src="./partials/services/smmInfo.html"></include>
     </main>
 
     <include src="./partials/footer.html"></include>
