+++ src/testImonials.html
@@ -15,6 +15,7 @@
     <script type="module" src="./index.js"></script>
   </head>
   <body>
+    <include src="./partials/header.html"></include>
     <main class="section" style="background-color: var(--background-pages)">
       <div class="container section">
         <include src="./partials/breadcrumbs.html"></include>
